Prevent error while loading missig includes
When someone fetches an object using .includes(:anything) but the server doesn't include 'anything' in its response (Because it might be null), fetching that missing include would result in an error.
To prevent this from happening a check is introduced: 'load_includes?' which prevents requests to try and load unavailable data.
To prevent this error from happening on collections, a check has been introduced to only map the response objects in case the parent objects defines a 'placeholder' for it.